+++ src/llamafactory/model/__init__.py
@@ -1,12 +1,12 @@
 from .loader import load_config, load_model, load_tokenizer
-from .utils.misc import find_all_linear_modules
-from .utils.valuehead import load_valuehead_params
+from .model_utils.misc import find_all_linear_modules
+from .model_utils.valuehead import load_valuehead_params
 
 
 __all__ = [
     "load_config",
     "load_model",
     "load_tokenizer",
-    "load_valuehead_params",
     "find_all_linear_modules",
+    "load_valuehead_params",
 ]
