+++ src/llamafactory/model/model_utils/visual.py
